putting general plot functions back in model.py
